Moved common external binary from url functionality to UrlBinary and made it the parent of LocalFileBinary.  Filled out functionality of UrlBinary
