+++ fcrepo-kernel-modeshape/src/main/java/org/fcrepo/kernel/modeshape/LocalFileBinary.java
@@ -17,23 +17,17 @@
  */
 package org.fcrepo.kernel.modeshape;
 
-import static org.modeshape.jcr.api.JcrConstants.JCR_CONTENT;
 import static org.modeshape.jcr.api.JcrConstants.JCR_DATA;
 import static org.slf4j.LoggerFactory.getLogger;
 
 import java.io.File;
-import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 import javax.jcr.Node;
-import javax.jcr.Property;
 import javax.jcr.RepositoryException;
 import org.apache.jena.rdf.model.Resource;
 import org.fcrepo.kernel.api.RdfStream;
@@ -44,11 +38,6 @@ import org.fcrepo.kernel.api.exception.UnsupportedAlgorithmException;
 import org.fcrepo.kernel.api.identifiers.IdentifierConverter;
 import org.fcrepo.kernel.api.models.FedoraResource;
 import org.fcrepo.kernel.api.services.policy.StoragePolicyDecisionPoint;
-import org.fcrepo.kernel.api.utils.CacheEntry;
-import org.fcrepo.kernel.api.utils.ContentDigest;
-import org.fcrepo.kernel.api.utils.FixityResult;
-import org.fcrepo.kernel.api.utils.MessageExternalBodyContentType;
-import org.fcrepo.kernel.modeshape.rdf.impl.FixityRdfContext;
 import org.fcrepo.kernel.modeshape.utils.FedoraTypesUtils;
 import org.fcrepo.kernel.modeshape.utils.impl.CacheEntryFactory;
 import org.slf4j.Logger;
@@ -61,7 +50,7 @@ import com.codahale.metrics.Timer;
  * @author bbpennel
  * @since 11/29/2017
  */
-public class LocalFileBinary extends AbstractFedoraBinary {
+public class LocalFileBinary extends UrlBinary {
     private static final Logger LOGGER = getLogger(LocalFileBinary.class);
 
     /**
@@ -71,19 +60,6 @@ public class LocalFileBinary extends AbstractFedoraBinary {
      */
     public LocalFileBinary(final Node node) {
         super(node);
-    }
-
-    /*
-     * (non-Javadoc)
-     * @see org.fcrepo.kernel.modeshape.FedoraBinaryImpl#getContent()
-     */
-    @Override
-    public InputStream getContent() {
-        try {
-            return getFileUri().toURL().openStream();
-        } catch (final IOException e) {
-            throw new RepositoryRuntimeException(e);
-        }
     }
 
     /*
@@ -124,7 +100,16 @@ public class LocalFileBinary extends AbstractFedoraBinary {
             final Collection<URI> nonNullChecksums = (null == checksums) ? new HashSet<>() : checksums;
             verifyChecksums(nonNullChecksums);
 
-            decorateLocalFileNode(contentNode, nonNullChecksums);
+            // Store checksums on node
+            final String[] checksumArray = new String[nonNullChecksums.size()];
+            nonNullChecksums.stream().map(Object::toString).collect(Collectors.toSet()).toArray(checksumArray);
+            contentNode.setProperty(CONTENT_DIGEST, checksumArray);
+
+            // Store the size of the file
+            final long size = getContentSize();
+            contentSizeHistogram.update(size);
+            contentNode.setProperty(CONTENT_SIZE, size);
+
             FedoraTypesUtils.touch(getNode());
             FedoraTypesUtils.touch(((FedoraResourceImpl) getDescription()).getNode());
 
@@ -135,39 +120,7 @@ public class LocalFileBinary extends AbstractFedoraBinary {
         }
     }
 
-    private void verifyChecksums(final Collection<URI> checksums)
-            throws InvalidChecksumException, RepositoryException {
 
-        final Property mimeTypeProperty = getProperty(HAS_MIME_TYPE);
-        final Map<URI, URI> checksumErrors = new HashMap<>();
-
-        // Loop through provided checksums validating against computed values
-        checksums.forEach(checksum -> {
-            final String algorithm = ContentDigest.getAlgorithm(checksum);
-            try {
-                final CacheEntry cacheEntry = CacheEntryFactory.forProperty(mimeTypeProperty);
-                cacheEntry.checkFixity(algorithm).stream().findFirst().ifPresent(
-                        fixityResult -> {
-                            if (!fixityResult.matches(checksum)) {
-                                LOGGER.debug("Failed checksum test");
-                                checksumErrors.put(checksum, fixityResult.getComputedChecksum());
-                            }
-                        }
-                );
-            } catch (final RepositoryException e) {
-                throw new RepositoryRuntimeException(e);
-            }
-        });
-
-        // Throw an exception if any checksum errors occurred
-        if (!checksumErrors.isEmpty()) {
-            final String template = "Checksum Mismatch of %1$s and %2$s\n";
-            final StringBuilder error = new StringBuilder();
-            checksumErrors.forEach((key, value) -> error.append(String.format(template, key, value)));
-            throw new InvalidChecksumException(error.toString());
-        }
-
-    }
 
     /*
      * (non-Javadoc)
@@ -179,7 +132,7 @@ public class LocalFileBinary extends AbstractFedoraBinary {
         if (sizeValue > -1L) {
             return sizeValue;
         }
-        final File file = new File(getFileUri().getPath());
+        final File file = new File(getResourceUri().getPath());
         return file.length();
     }
 
@@ -191,35 +144,6 @@ public class LocalFileBinary extends AbstractFedoraBinary {
     @Override
     public RdfStream getFixity(final IdentifierConverter<Resource, FedoraResource> idTranslator) {
         return getFixity(idTranslator, getContentDigest(), getContentSize());
-    }
-
-    /*
-     * (non-Javadoc)
-     * @see
-     * org.fcrepo.kernel.modeshape.FedoraBinaryImpl#getFixity(org.fcrepo.kernel.api.identifiers.IdentifierConverter,
-     * java.net.URI, long)
-     */
-    @Override
-    public RdfStream getFixity(final IdentifierConverter<Resource, FedoraResource> idTranslator, final URI digestUri,
-            final long size) {
-
-        fixityCheckCounter.inc();
-
-        try (final Timer.Context context = timer.time()) {
-
-            LOGGER.debug("Checking resource: " + getPath());
-
-            final String algorithm = ContentDigest.getAlgorithm(digestUri);
-
-            final long contentSize = size < 0 ? getContentSize() : size;
-
-            final Collection<FixityResult> fixityResults
-                    = CacheEntryFactory.forProperty(getProperty(HAS_MIME_TYPE)).checkFixity(algorithm);
-
-            return new FixityRdfContext(this, idTranslator, fixityResults, digestUri, contentSize);
-        } catch (final RepositoryException e) {
-            throw new RepositoryRuntimeException(e);
-        }
     }
 
     /*
@@ -246,69 +170,6 @@ public class LocalFileBinary extends AbstractFedoraBinary {
 
             return CacheEntryFactory.forProperty(getProperty(HAS_MIME_TYPE)).checkFixity(algorithms);
         } catch (final RepositoryException e) {
-            throw new RepositoryRuntimeException(e);
-        }
-    }
-
-    private String getResourceLocation() {
-        try {
-            final MessageExternalBodyContentType externalBody
-                    = MessageExternalBodyContentType.parse(super.getMimeType());
-            return externalBody.getResourceLocation();
-        } catch (final UnsupportedAccessTypeException e) {
-            throw new RepositoryRuntimeException(e);
-        }
-    }
-
-    private URI getFileUri() {
-        try {
-            return new URI(getResourceLocation());
-        } catch (final URISyntaxException e) {
-            throw new RepositoryRuntimeException(e);
-        }
-    }
-
-    private void decorateLocalFileNode(final Node contentNode, final Collection<URI> checksums)
-            throws RepositoryException {
-        if (contentNode == null) {
-            LOGGER.warn("{} node appears to be null!", JCR_CONTENT);
-            return;
-        }
-
-        final long size = getContentSize();
-        contentSizeHistogram.update(size);
-
-        // checksums.add(ContentDigest.asURI(SHA1.algorithm, dsChecksum));
-
-        final String[] checksumArray = new String[checksums.size()];
-        checksums.stream().map(Object::toString).collect(Collectors.toSet()).toArray(checksumArray);
-
-        contentNode.setProperty(CONTENT_DIGEST, checksumArray);
-        contentNode.setProperty(CONTENT_SIZE, size);
-
-        LOGGER.debug("Decorated data property at path: {}", getResourceLocation());
-    }
-
-    /**
-     * Returns the specified mimetype in place of the original external-body if provided
-     */
-    @Override
-    public String getMimeType() {
-        final String mimeType = super.getMimeType();
-        if (!MessageExternalBodyContentType.isExternalBodyType(mimeType)) {
-            // Non-external-body mimetype, return without further processing
-            return mimeType;
-        }
-        try {
-            final MessageExternalBodyContentType extBodyType = MessageExternalBodyContentType.parse(mimeType);
-            // Return the overridden mimetype if one is available, otherwise give the original content type
-            final String mimeTypeOverride = extBodyType.getMimeType();
-            if (mimeTypeOverride == null) {
-                return mimeType;
-            } else {
-                return mimeTypeOverride;
-            }
-        } catch (final UnsupportedAccessTypeException e) {
             throw new RepositoryRuntimeException(e);
         }
     }
